+++ src/setup/internal.ts
@@ -5,11 +5,25 @@ export type Endpoints = {
   user: string
 }
 
+export type UserMessages = {
+  versionMismatch: {
+    newer(version: string): Promise<void>
+    older(version: string): Promise<void>
+  }
+}
+
+type Setup = {
+  debug: boolean
+  endpoints: Endpoints
+  userMessages: UserMessages
+  shouldPin: boolean
+}
+
 
 /**
  * @internal
  */
-export const setup = {
+export const setup: Setup = {
   debug: false,
 
   endpoints: {
@@ -19,5 +33,16 @@ export const setup = {
     user: "fission.name"
   },
 
+  userMessages: {
+    versionMismatch: {
+      newer: async () => alertIfPossible(`Sorry, we can't sync your filesystem with this app. This app only understands older versions of filesystems. Please try to hard refresh this site or let this app's developer know. Feel free to contact Fission support: support@fission.codes`),
+      older: async () => alertIfPossible(`Sorry, we can't sync your filesystem with this app. Your filesystem version is out-dated and it needs to be migrated. Use the migration app or talk to Fission support: support@fission.codes`),
+    }
+  },
+
   shouldPin: false,
+}
+
+function alertIfPossible(str: string) {
+  if (globalThis.alert != null) globalThis.alert(str)
 }