+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.29.1
 
 - Check the wnfs version field when initialising a filesystem and alert users about outdated filesystems or outdated apps.
+- Make version mismatch errors configurable through `setup.userMessages(...)`.
 
 
 ### v0.29.0